Arreglos en fotos

+++ server/server.js
@@ -5,7 +5,7 @@ const http = require('http');
 const server = http.createServer(app);
 const { Server } = require("socket.io");
 const io = new Server(server);
-const { writeFile, unlink } = require("fs");
+const { writeFile, unlink, existsSync } = require("fs");
 const path = require("path");
 
 // Para hacerlo con framework:
@@ -52,7 +52,7 @@ io.on('connection', (socket) => {
     socket.userID = id;
   });
 
-  console.log(`CONECTADO --> ID: ${socket.userID}`);
+  console.log(`CONECTADO --> ID: ${socket.id}`);
 
   socket.connectedRoomUsers = 0;
 
@@ -161,13 +161,20 @@ io.on('connection', (socket) => {
     io.to(socket.room).emit('typing', {userID: data.userID, typing: data.typing});
   });
 
+  socket.on('getPhotoOfUser', (data) => {
+    socket.emit('photoOfUser', {
+      path: `userPhotos/${data.userID}.jpg`,
+      userID: data.userID,
+      username: data.username
+    });
+  });
 
   // Desconectar usuario
   socket.on('disconnect', () => {
     switch(socket.room){
       case '1':
         connectedUsersRoom1--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom1);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom1});
         usersRoom1.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom1.splice(index, 1);
@@ -176,7 +183,7 @@ io.on('connection', (socket) => {
         break;
       case '2':
         connectedUsersRoom2--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom2);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom2});
         usersRoom2.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom2.splice(index, 1);
@@ -185,7 +192,7 @@ io.on('connection', (socket) => {
         break;
       case '3':
         connectedUsersRoom3--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom3);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom3});
         usersRoom3.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom3.splice(index, 1);
@@ -194,7 +201,7 @@ io.on('connection', (socket) => {
         break;
       case '4':
         connectedUsersRoom4--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom4);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom4});
         usersRoom4.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom4.splice(index, 1);
@@ -203,7 +210,7 @@ io.on('connection', (socket) => {
         break;
       case '5':
         connectedUsersRoom5--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom5);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom5});
         usersRoom5.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom5.splice(index, 1);
@@ -212,7 +219,7 @@ io.on('connection', (socket) => {
         break;
       case '6':
         connectedUsersRoom6--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom6);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom6});
         usersRoom6.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom6.splice(index, 1);
@@ -221,7 +228,7 @@ io.on('connection', (socket) => {
         break;
       case '7':
         connectedUsersRoom7--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom7);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom7});
         usersRoom7.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom7.splice(index, 1);
@@ -230,7 +237,7 @@ io.on('connection', (socket) => {
         break;
       case '8':
         connectedUsersRoom8--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom8);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom8});
         usersRoom8.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom8.splice(index, 1);
@@ -239,7 +246,7 @@ io.on('connection', (socket) => {
         break;
       case '9':
         connectedUsersRoom9--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom9);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom9});
         usersRoom9.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom9.splice(index, 1);
@@ -248,7 +255,7 @@ io.on('connection', (socket) => {
         break;
       case '10':
         connectedUsersRoom10--;
-        io.to(socket.room).emit('usersConnected', connectedUsersRoom10);
+        io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username, usersConnected: connectedUsersRoom10});
         usersRoom10.forEach((user, index) => {
           if(user.id == socket.userID){
             usersRoom10.splice(index, 1);
@@ -257,15 +264,16 @@ io.on('connection', (socket) => {
         break;
     }
 
-    io.to(socket.room).emit('userDisconnected', {id: socket.userID, username: socket.username});
     console.log(`DESCONECTADO --> ID: ${socket.userID}`);
 
-    unlink(`${socket.pathToUpload}/${socket.userID}.jpg`, (err) => {
-      if(err){
-        console.log(err);
+    if(existsSync(`${socket.pathToUpload}/${socket.userID}.jpg`) == true){
+      unlink(`${socket.pathToUpload}/${socket.userID}.jpg`, (err) => {
+        if(err){
+          console.log(err);
+        }
       }
+      );
     }
-    );
 
   });
 });
